Passing incomplete chains up from GPU to tiler [#6]
When all the work was completed on the GPU, it was possible to simply start the chains from the highest scoring cell. Unfortunately, this doesn't actually work when we are calculating seperate tiles. This ensures that chains that trail off the edge of the tile are allowed to be passed on to the next tile